// src/payload-types.ts
export interface ConsolesMedia {
  id: number;
  alt?: string | null;
  /**
   * Chemin calculé automatiquement depuis les relations
   */
  mediaPath?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "games-media".
 */
export interface GamesMedia {
  id: number;
  alt?: string | null;
  /**
   * Chemin calculé automatiquement depuis les relations
   */
  mediaPath?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Console {
  id: number;
  name: string;
  slug: string;
  /**
   * Nom de l'icône (ex: switch, 3ds, gba...)
   */
  icon?: string | null;
  image?: (number | null) | ConsolesMedia;
  releaseDate?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "games".
 */
export interface Game {
  id: number;
  title: string;
  slug: string;
  console: number | Console;
  /**
   * Nom de l'icône (ex: mario, zelda...)
   */
  icon?: string | null;
  description?: string | null;
  coverImage?: (number | null) | GamesMedia;
  releaseDate?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "categories".
 */
export interface Category {
  id: number;
  name: string;
  game: number | Game;
  updatedAt: string;
  createdAt: string;
}